rework to prep for pyinstaller build
large-scale rework to prep for pyinstaller build

update module imports
