+++ dependencies.md
@@ -1,0 +1,19 @@
+## Dependencies
+This list of required commands and library names, plus their APT package names in Ubuntu 20.04.
+
+### binaries
+
+- ldconfig: libc-bin (should be in every system)
+- wget: wget
+- wine: fuse (AppImage), winbind
+- winetricks: winetricks, binutils, cabextract, wget
+
+### system libraries
+
+- libfuse (libfuse2, dependency of 'fuse')
+
+### python libraries
+
+- Standard Library (libpython3-stdlib)
+- Tk (python3-tk)
+- psutil (python3-psutil)
